BREAKING again, since even more information will be needed
If we're breaking things, might as well break them to the end! I'll need
to get exactly what changed in EffectNotify so I could know how many
particles to spawn. But not right now. So for now this is just senseless
breaking, but so that I would avoid breaking things in the future. Treat
the last parameter to EffectNotify as "reserved".

Signed-off-by: GreatEmerald <pastas4@gmail.com>